Add deep copy of hatches/eggs
Add deep copy of hatches and eggs when we save to the cache file. This prevents the time values from getting turned into strings when writing to the cache. Time values getting converted to strings was causing an error when the values were not loaded back to datetime due to reduced loadCache() calls. The strings would get compared to datetimes and used as datetimes and it would cause problems.